Move test_put_* from test_clib to test_clib_put
